``` feat(function): add NewInstance method to invoke constructor functions
Add a new method NewInstance to the Function type that allows invoking
constructor functions to create object instances. This method takes
variadic Valuer arguments and returns an Object pointer or an error.

The implementation converts Go Valuer arguments to C value pointers,
calls the C function FunctionNewInstance, and wraps the result in a
Go Object.
```